+++ src/ImageParsers/PdfJpgParser.php
@@ -33,16 +33,17 @@ class PdfJpgParser implements PdfImageParserInterface
         /* @phpstan-var array{0: int, 1: int, 2: int, channels?: int, bits?: int}|false $size */
         $size = \getimagesize($file);
         if (!\is_array($size)) {
-            throw PdfException::format('Missing or incorrect image file: %s.', $file);
+            throw PdfException::format('Missing or invalid image size: %s.', $file);
         }
 
         if (\IMG_JPG !== $size[2]) {
-            throw PdfException::format('The file is not a JPEG image: %s.', $file);
+            // Interlacing %d not supported: %s.
+            throw PdfException::format('Invalid JPEG image type (%d): %s.', $size[2], $file);
         }
 
         /** @phpstan-var int<3,5> $channels */
         $channels = $size['channels'] ?? 3;
-        $color_space = match ($channels) {
+        $colorSpace = match ($channels) {
             3 => 'DeviceRGB',
             4 => 'DeviceCMYK',
             default => 'DeviceGray'
@@ -55,7 +56,7 @@ class PdfJpgParser implements PdfImageParserInterface
             'number' => 0,
             'width' => $size[0],
             'height' => $size[1],
-            'color_space' => $color_space,
+            'color_space' => $colorSpace,
             'bits_per_component' => $bitsPerComponent,
             'filter' => 'DCTDecode',
             'data' => $data,