+++ tests/ImageParsers/PdfGifParserTest.php
@@ -21,6 +21,7 @@ class PdfGifParserTest extends AbstractPdfParserTestCase
     public function testInvalid(): void
     {
         self::expectException(PdfException::class);
+        self::expectExceptionMessageMatches('/Missing or incorrect image file:.*image.fake.$/');
         $file = 'image.fake';
         $this->parseFile($file);
     }