+++ tests/ImageParsers/PdfJpgParserTest.php
@@ -37,6 +37,7 @@ class PdfJpgParserTest extends AbstractPdfParserTestCase
     public function testInvalid(): void
     {
         self::expectException(PdfException::class);
+        self::expectExceptionMessageMatches('/Missing or invalid image size:.*image.fake.$/');
         $file = 'image.fake';
         $this->parseFile($file);
     }
@@ -44,6 +45,7 @@ class PdfJpgParserTest extends AbstractPdfParserTestCase
     public function testInvalidType(): void
     {
         self::expectException(PdfException::class);
+        self::expectExceptionMessageMatches('/Invalid JPEG image type \(1\):.*image.gif.$/');
         $file = 'image.gif';
         $this->parseFile($file);
     }