+++ tests/ImageParsers/PdfPngParserTest.php
@@ -41,6 +41,7 @@ class PdfPngParserTest extends AbstractPdfParserTestCase
     public function testEndOfStream(): void
     {
         self::expectException(PdfException::class);
+        self::expectExceptionMessage('Unexpected end of stream.');
         $file = 'image_truncate.png';
         $this->parseFile($file);
     }
@@ -48,6 +49,7 @@ class PdfPngParserTest extends AbstractPdfParserTestCase
     public function testInvalid(): void
     {
         self::expectException(PdfException::class);
+        self::expectExceptionMessageMatches('/Unable to open image file:.*image.fake.$/');
         $file = 'image.fake';
         $this->parseFile($file);
     }
@@ -55,6 +57,7 @@ class PdfPngParserTest extends AbstractPdfParserTestCase
     public function testInvalidBpc(): void
     {
         self::expectException(PdfException::class);
+        self::expectExceptionMessageMatches('/Bits per component 16 not supported:.*invalid_bpc.png.$/');
         $file = 'invalid_bpc.png';
         $this->parseFile($file);
     }
@@ -62,6 +65,7 @@ class PdfPngParserTest extends AbstractPdfParserTestCase
     public function testInvalidColorType(): void
     {
         self::expectException(PdfException::class);
+        self::expectExceptionMessageMatches('/Color type 12 not supported:.*invalid_color_type.png.$/');
         $file = 'invalid_color_type.png';
         $this->parseFile($file);
     }
@@ -69,6 +73,7 @@ class PdfPngParserTest extends AbstractPdfParserTestCase
     public function testInvalidColorTypeWithPalette(): void
     {
         self::expectException(PdfException::class);
+        self::expectExceptionMessageMatches('/Missing palette:.*invalid_color_type_with_palette.png.$/');
         $file = 'invalid_color_type_with_palette.png';
         $this->parseFile($file);
     }
@@ -76,6 +81,7 @@ class PdfPngParserTest extends AbstractPdfParserTestCase
     public function testInvalidCompression(): void
     {
         self::expectException(PdfException::class);
+        self::expectExceptionMessageMatches('/Compression method 16 not supported:.*invalid_compression.png.$/');
         $file = 'invalid_compression.png';
         $this->parseFile($file);
     }
@@ -83,6 +89,7 @@ class PdfPngParserTest extends AbstractPdfParserTestCase
     public function testInvalidFilter(): void
     {
         self::expectException(PdfException::class);
+        self::expectExceptionMessageMatches('/Filter method 16 not supported:.*invalid_filter.png.$/');
         $file = 'invalid_filter.png';
         $this->parseFile($file);
     }
@@ -90,6 +97,7 @@ class PdfPngParserTest extends AbstractPdfParserTestCase
     public function testInvalidHeaderChunk(): void
     {
         self::expectException(PdfException::class);
+        self::expectExceptionMessageMatches('/Incorrect PNG header chunk \(.*\):.*invalid_header_chunk.png.$/');
         $file = 'invalid_header_chunk.png';
         $this->parseFile($file);
     }
@@ -97,6 +105,7 @@ class PdfPngParserTest extends AbstractPdfParserTestCase
     public function testInvalidHeaderLength(): void
     {
         self::expectException(PdfException::class);
+        self::expectExceptionMessageMatches('/Incorrect PNG header length \(0\):.*invalid_header_length.png.$/');
         $file = 'invalid_header_length.png';
         $this->parseFile($file);
     }
@@ -104,6 +113,7 @@ class PdfPngParserTest extends AbstractPdfParserTestCase
     public function testInvalidInterlacing(): void
     {
         self::expectException(PdfException::class);
+        self::expectExceptionMessageMatches('/Interlacing 16 not supported:.*invalid_interlacing.png.$/');
         $file = 'invalid_interlacing.png';
         $this->parseFile($file);
     }
@@ -111,6 +121,7 @@ class PdfPngParserTest extends AbstractPdfParserTestCase
     public function testInvalidSignature(): void
     {
         self::expectException(PdfException::class);
+        self::expectExceptionMessageMatches('/Incorrect PNG header signature:.*invalid_signature.png.$/');
         $file = 'invalid_signature.png';
         $this->parseFile($file);
     }