+++ tests/PdfDocCellTest.php
@@ -53,6 +53,7 @@ class PdfDocCellTest extends AbstractPdfDocTestCase
     public function testWithoutFont(): void
     {
         self::expectException(PdfException::class);
+        self::expectExceptionMessage('No font is set.');
         $doc = $this->createDocument(true, false);
         $doc->cell(text: 'fake');
     }