+++ tests/Traits/PdfBookmarkTraitTest.php
@@ -73,6 +73,7 @@ class PdfBookmarkTraitTest extends TestCase
     public function testLevelNegative(): void
     {
         self::expectException(PdfException::class);
+        self::expectExceptionMessage('Invalid bookmark level: -1. Allowed value: 0.');
         $doc = $this->createDocument();
         // @phpstan-ignore argument.type
         $doc->addBookmark('Negative Level', level: -1);